+++ next.config.js
@@ -2,6 +2,21 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
+  compiler: {
+    reactRemoveProperties: true,
+  },
+  i18n: {
+    locales: ['en'],
+    defaultLocale: 'en'
+  },
+  async rewrites() {
+    return [
+      {
+        source: '/',
+        destination: '/topics/react',
+      },
+    ];
+  },
 }
 
 module.exports = nextConfig